\name{EMMIXuskew-package}
\alias{EMMIXuskew-package}
\alias{EMMIXuskew}
\docType{package}
\title{
Finite mixture of unrestricted multivariate skew t distributions  
}
\description{
Theis package implements an EM algorithm for 
fitting mixtures of unrestricted multivariate skew t (FM-uMST) distributions.
Functions for random sample generation, discriminant analysis, 
and visualisation (in 2D and 3D) is also provided.
}
\details{
\tabular{ll}{
Package: \tab EMMIXuskew\cr
Type: \tab Package\cr
Version: \tab 0.11-5\cr
Date: \tab 2013-11-11\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
S.X. Lee, G.J. McLachlan
}
\references{
Lee S, McLachlan G (2011). On the fitting of mixtures of multivariate skew t-distributions
via the EM algorithm. arXiv:1109.4706 [stat.ME]

Lee, S. and McLachlan, G.J. (2012) Finite mixtures of multivariate 
skew t-distributions: some recent and new results. 
Statistics and Computing. To appear
                                     
Lee, S. and McLachlan, G.J. (2013) EMMIXuskew: An \code{R} package for
fitting mixtures of multivariate skew t-distributions via the EM algorithm.
Journal of Statistical Software, 55(12), 1-22.
URL http://www.jstatsoft.org/v55/i12/.
}
\keyword{ package }
\seealso{
\code{\link{fmmst}}, \code{\link{dfmmst}}, \code{\link{rfmmst}}, \code{\link{fmmst.contour.3d}}
}
